+++ pixelbin/platform/models/AddPresetResponse.py
@@ -14,6 +14,14 @@ from ..models.BaseSchema import BaseSchema
 
 
 
+
+
+
+
+
+
+
+
 class AddPresetResponse(BaseSchema):
     # Assets swagger.json
 
@@ -25,5 +33,13 @@ class AddPresetResponse(BaseSchema):
     params = fields.Dict(required=False)
     
     archived = fields.Boolean(required=False)
+    
+    orgId = fields.Int(required=False)
+    
+    isActive = fields.Boolean(required=False)
+    
+    createdAt = fields.Str(required=False)
+    
+    updatedAt = fields.Str(required=False)